+++ tests/k8s-manager.spec.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import Environment from '../lib/env-manager';
-import K8sManager from '../lib/k8s-manager';
-import { IUpgradeMessage } from '../lib/upgrade-message';
-import { runCommand } from '../utils/command-exec';
+import Environment from '../src/lib/env-manager';
+import K8sManager from '../src/lib/k8s-manager';
+import { IUpgradeMessage } from '../src/lib/upgrade-message';
+import { runCommand } from '../src/utils/command-exec';
 import { k8s_deployment_name, tempNamespace } from './resources/test-constants';
 import { expect } from 'chai';
 import { before, beforeEach } from 'mocha';
@@ -15,10 +15,10 @@ describe('k8s-manager', () => {
 
   before(async () => {
     await runCommand(
-      `kubectl apply -f src/__tests__/resources/nginx.default.yaml`,
+      `kubectl apply -f tests/resources/nginx.default.yaml`,
       'Creating an nginx deployment in the default namespace');
     await runCommand(
-      `kubectl -n ${tempNamespace} apply -f src/__tests__/resources/nginx.yaml`, 'Creating an nginx deployment');
+      `kubectl -n ${tempNamespace} apply -f tests/resources/nginx.yaml`, 'Creating an nginx deployment');
     await runCommand(`sleep 2`, 'Waiting a few seconds...');
   });
