+++ tests/resources/nginx.default.yaml
