+++ tests/resources/nginx.yaml
