+++ tests/upgrade-service.spec.ts
@@ -1,8 +1,8 @@
-import { IUpgradeMessage } from '../lib/upgrade-message';
-import { runCommand } from '../utils/command-exec';
-import UpgradeService from '../lib/upgrade-service';
+import { IUpgradeMessage } from '../src/lib/upgrade-message';
+import { runCommand } from '../src/utils/command-exec';
+import UpgradeService from '../src/lib/upgrade-service';
 import { k8s_deployment_name, tempNamespace } from './resources/test-constants';
-import { UpgradeResult } from '../lib/upgrade-result';
+import { UpgradeResult } from '../src/lib/upgrade-result';
 import { before } from 'mocha';
 import { expect } from 'chai';
 import sinon from 'sinon';
@@ -13,7 +13,7 @@ describe('Upgrade Service', () => {
 
   before(async () => {
     await runCommand(
-      `kubectl -n ${tempNamespace} apply -f src/__tests__/resources/busybox.yaml`,
+      `kubectl -n ${tempNamespace} apply -f tests/resources/busybox.yaml`,
       'Creating a busybox deployment');
     await runCommand(`sleep 2`, 'Waiting a few seconds...');
   });
